Fix for Domains Location and Compability Changes
Changed the domains location to default to the static entry in the
config file if there is no domains file found in the adjacent or
contained folders from where the script is run.
Also changed the NG911_DataChecks functions to use a class to get the
information for path locations. This makes it easier to pass information
into them when they are called from the GUI.